% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aiDetectR_app.R
\name{run_aiDetectR_app}
\alias{run_aiDetectR_app}
\title{Launch the aiDetectR Shiny App}
\usage{
run_aiDetectR_app()
}
\value{
Opens a Shiny app in the browser.
}
\description{
Starts the interactive AI vs Human image detection dashboard.
}
\examples{
if (interactive()) {
  run_aiDetectR_app()
}
}
