% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_imgdet_env.R
\name{setup_imgdet_env}
\alias{setup_imgdet_env}
\title{Set up the Python environment for aiDetectR}
\usage{
setup_imgdet_env(env = "imgdet", python_version = "3.11")
}
\arguments{
\item{env}{The name of the conda environment to create (default is "imgdet")}

\item{python_version}{Python version to install (default is "3.11")}
}
\value{
NULL (called for side effect)
}
\description{
This installs the required Python packages into a named conda environment ("imgdet").
Run this once after installing the package.
}
